fix: Move code from client.rs to client/player.rs
